added the dashboard file to test the code
you need to create a database on your local machine in order to test this. The name of my database is "advising" and the student table's name is "students". In my database, there are 7 columns. to test this code, you just need the student_id, user_name and the email.